Restore the ScrollFix hack.
One thing that is obviously better is dealing with the cursor change --
it didn't work on elements that have their own cursor, and it risks
restoring the body cursor to "auto" when it wasn't that, and outside of
the body on smaller-than-the-screen pages.  But it's probably good for
other things like grabbing events etc.